requirements.txt
pytest==8.3.5
pycryptodome==3.21.0